% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_def}
\alias{get_def}
\title{Obtain definitions and its POS of an English word.}
\usage{
get_def(voc)
}
\arguments{
\item{voc}{Original English word. The definition of \code{voc} will be scraped}
}
\value{
Definitions and its POS of an English word \code{voc} will be generated.
}
\description{
Obtain definitions and its POS of an English word.
}
\examples{
get_def("tuna")
}
