% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_syn}
\alias{get_syn}
\title{Obtain synonyms and its POS of an English word.}
\usage{
get_syn(voc)
}
\arguments{
\item{voc}{Original English word. The synonyms of \code{voc} will be scraped}
}
\value{
Synonyms and its POS of an English word \code{voc} will be generated.
}
\description{
Obtain synonyms and its POS of an English word.
}
\examples{
get_syn("great")
}
